removed old v2 fields from user, added progress model class

+++ Services.Habitica/Model/Progress.cs
@@ -19,9 +19,9 @@ namespace Trellabit.Services.Habitica.Model
 		public int CollectedItems { get; set; }
 
 		[JsonProperty("down")]
-		public float Down { get; set; }
+		public double Down { get; set; }
 
 		[JsonProperty("up")]
-		public float Up { get; set; }
+		public double Up { get; set; }
 	}
 }
